# openapi.json
{
  "components": {
    "schemas": {
...
      "ClipList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SampleClip"
            },
            "title": "Items",
            "type": "array"
          },
          "provider": {
            "description": "Active voice provider.",
            "examples": [
              "sample"
            ],
            "title": "Provider",
            "type": "string"
          }
        },
        "required": [
          "provider",
          "items"
        ],
        "title": "ClipList",
        "type": "object"
      },
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "description": "Stable machine-readable code, e.g. `not_found`.",
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional structured context, e.g. field validation errors.",
            "title": "Detail"
          },
          "message": {
            "description": "Human-readable message. Safe to surface in UI.",
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "detail"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
...
      "ErrorResponse": {
        "description": "Every non-2xx response from `/api` has this shape.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "SampleClip": {
        "properties": {
          "duration_ms": {
            "title": "Duration Ms",
            "type": "integer"
          },
          "id": {
            "examples": [
              "greeting"
            ],
            "title": "Id",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "id",
          "text",
          "duration_ms"
        ],
        "title": "SampleClip",
        "type": "object"
      },
...
      "SpeakRandomRequest": {
        "description": "Play random sample clips \u2014 the smoke test for the whole audio-output path.",
        "properties": {
          "count": {
            "default": 1,
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Count",
            "type": "integer"
          },
          "gap_seconds": {
            "default": 0.0,
            "description": "Silence between clips, on top of each clip's own length.",
            "maximum": 60.0,
            "minimum": 0.0,
            "title": "Gap Seconds",
            "type": "number"
          }
        },
        "title": "SpeakRandomRequest",
        "type": "object"
      },
      "SpeakRequest": {
        "description": "Ask Meet AGI to say something out loud in a meeting.\n\nExactly one of `text` or `clip_id`. `text` goes through the voice provider;\n`clip_id` plays a known sample clip verbatim, which is what the demo path uses and\nwhat makes a stage failure recoverable.",
        "properties": {
          "clip_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "A sample clip id. `GET /api/speech/clips` lists them.",
            "examples": [
              "greeting"
            ],
            "title": "Clip Id"
          },
          "text": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          }
        },
        "title": "SpeakRequest",
        "type": "object"
      },
...
      "Utterance": {
        "properties": {
          "clip_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Set when the audio came from a fixed sample clip rather than TTS.",
            "examples": [
              "flag_revenue"
            ],
            "title": "Clip Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "duration_ms": {
            "default": 0,
            "title": "Duration Ms",
            "type": "integer"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "id": {
            "examples": [
              "utt_01J8XP6N7P8Q9R0S1T2U3V"
            ],
            "title": "Id",
            "type": "string"
          },
          "meeting_id": {
            "title": "Meeting Id",
            "type": "string"
          },
          "placeholder": {
            "default": false,
            "description": "True when the audio is a stand-in that does not say `requested_text`. The UI should label it rather than presenting it as real speech.",
            "title": "Placeholder",
            "type": "boolean"
          },
          "played_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Played At"
          },
          "requested_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "What the caller asked Meet AGI to say, when that differs from `text`. Non-null only while the voice provider is a placeholder.",
            "title": "Requested Text"
          },
          "status": {
            "$ref": "#/components/schemas/UtteranceStatus",
            "default": "queued"
          },
          "text": {
            "description": "What the audio actually says.",
            "examples": [
              "Quick flag. That revenue number conflicts with the Q3 board deck."
            ],
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "id",
          "meeting_id",
          "text",
          "requested_text",
          "clip_id",
          "placeholder",
          "duration_ms",
          "status",
          "error",
          "created_at",
          "played_at"
        ],
        "title": "Utterance",
        "type": "object"
      },
      "UtteranceStatus": {
        "enum": [
          "queued",
          "speaking",
          "played",
          "dropped",
          "failed"
        ],
        "title": "UtteranceStatus",
        "type": "string"
      },
...
    }
  },
  "info": {
    "description": "\nBackend for **Meet AGI**, a Google Meet copilot that actually speaks.\n\n### Building the frontend against this\n\nYou do not need a Google Meet, a Recall.ai key, or an internet connection. Start a\nfixture replay and subscribe to the socket:\n\n```bash\ncurl -X POST localhost:8000/api/dev/harness/start \\\n  -H 'content-type: application/json' \\\n  -d '{\"fixture_id\":\"q3_revenue_review\",\"speed\":6,\"loop\":true}'\n```\n\nThat emits the complete real-time event stream \u2014 participants, partial and final\ntranscript, speaking states, wake-word detection, agent state transitions, and\ninterjections with real citations. It is byte-for-byte the same contract a real meeting\nproduces; only `Meeting.source` differs.\n\n### Contract rules\n\n- Types are generated from this OpenAPI document. Never hand-write an API type.\n- All mutations are REST. The WebSocket is server-to-client only.\n- Adding fields, enum variants, and event types is free and unannounced. Renames and\n  removals get a heads-up first.\n- `switch` on `LiveEvent.type` with a `default` branch. New event types will arrive.\n",
    "title": "Meet AGI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/meetings/{meeting_id}/interrupt": {
      "post": {
        "description": "The dashboard's equivalent of saying **\"AGI, stop talking\"** out loud. Drops the speech queue, cancels the reasoning that was about to produce more speech, and retracts the clip currently playing through Recall's stop-audio endpoint \u2014 so this really does stop it mid-word, not at the end of the sentence.\n\nReturns the utterances that were dropped, the one that was playing included.",
        "operationId": "interrupt_api_meetings__meeting_id__interrupt_post",
        "parameters": [
          {
            "in": "path",
            "name": "meeting_id",
            "required": true,
            "schema": {
              "title": "Meeting Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Utterance"
                  },
                  "title": "Response Interrupt Api Meetings  Meeting Id  Interrupt Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "Cut Meet AGI off mid-sentence",
        "tags": [
          "speech"
        ]
      }
    },
...
    "/api/meetings/{meeting_id}/speak": {
      "post": {
        "description": "Queues audio and returns immediately with the utterance in `queued`. Playback is serialized per meeting, so this is safe to call several times in a row \u2014 the clips play in order rather than on top of each other.\n\nPass `text` to go through the voice provider, or `clip_id` to play a known sample clip verbatim. While the voice provider is a placeholder, `text` still produces a canned clip; the utterance is flagged `placeholder: true` and keeps the requested text in `requested_text` so the UI can be honest about it.",
        "operationId": "speak_api_meetings__meeting_id__speak_post",
        "parameters": [
          {
            "in": "path",
            "name": "meeting_id",
            "required": true,
            "schema": {
              "title": "Meeting Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpeakRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Utterance"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "Say something in the meeting",
        "tags": [
          "speech"
        ]
      }
    },
    "/api/meetings/{meeting_id}/speak/random": {
      "post": {
        "description": "Smoke test for the audio path: queues `count` random sample clips, never the same one twice in a row. Use it to confirm the bot can be heard before trusting anything upstream of it.",
        "operationId": "speak_random_api_meetings__meeting_id__speak_random_post",
        "parameters": [
          {
            "in": "path",
            "name": "meeting_id",
            "required": true,
            "schema": {
              "title": "Meeting Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/SpeakRandomRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Utterance"
                  },
                  "title": "Response Speak Random Api Meetings  Meeting Id  Speak Random Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "Play random sample clips",
        "tags": [
          "speech"
        ]
      }
    },
...
    "/api/meetings/{meeting_id}/utterances": {
      "get": {
        "description": "Oldest first. Records are live \u2014 an utterance moves through `queued \u2192 speaking \u2192 played` in place, so polling this shows progress.",
        "operationId": "list_utterances_api_meetings__meeting_id__utterances_get",
        "parameters": [
          {
            "in": "path",
            "name": "meeting_id",
            "required": true,
            "schema": {
              "title": "Meeting Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Utterance"
                  },
                  "title": "Response List Utterances Api Meetings  Meeting Id  Utterances Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "What Meet AGI has said in this meeting",
        "tags": [
          "speech"
        ]
      }
    },
...
    "/api/speech/clips": {
      "get": {
        "description": "The stand-in voice while TTS is not wired up. These ids are what `clip_id` accepts on `/speak`, and the source of the audio `/speak/random` plays.",
        "operationId": "list_clips_api_speech_clips_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClipList"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "List the sample clips Meet AGI can play",
        "tags": [
          "speech"
        ]
      }
    }
  }
}